Factored out the project_activations function that is bottlenecking me; I will refactor this guy for mem efficiency.
