+++ sparse_coding/act_config.yaml
@@ -1,6 +1,6 @@
 # Model
-MODEL_DIR: "meta-llama/Llama-2-70b-hf"
-ACTS_LAYER: 32
+MODEL_DIR: "EleutherAI/pythia-70m"
+ACTS_LAYER: 2
 
 # Encoder Size
 PROJECTION_FACTOR: 10
@@ -13,9 +13,9 @@ BIASES_PATH: "acts_data/learned_biases.pt"
 TOP_K_INFO_PATH: "acts_data/token_info.csv"
 
 # Autoencoder Training
-LAMBDA_L1: 3
-LEARNING_RATE: 1.0e-3
-NUM_WORKERS: 16
+LAMBDA_L1: 1e-2
+LEARNING_RATE: 1e-2
+NUM_WORKERS: 4
 
 # Reproducibility
 SEED: 0
